refact: Order form logic separated down to component
